fix: compilation on MSVC

+++ src/utils/read_file_to_string.cpp
@@ -10,10 +10,10 @@
 auto engine::read_file_to_string(const std::filesystem::path& filename) -> std::optional<std::string> {
     std::ifstream ifs(filename.c_str(), std::ios::in | std::ios::binary | std::ios::ate);
 
-    ssize_t const fileSize = static_cast<size_t>(ifs.tellg());
+    size_t const fileSize = static_cast<size_t>(ifs.tellg());
     ifs.seekg(0, std::ios::beg);
 
-    if (fileSize == -1) {
+    if (fileSize == static_cast<size_t>(-1)) {
         return std::nullopt;
     }
